requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai==1.3.5
anthropic==0.9.0
playwright==1.40.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
pyyaml==6.0.1
pytest==7.4.0
pillow==10.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
werkzeug==2.3.7
jinja2==3.1.2
markupsafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
colorama==0.4.6
typing-extensions==4.7.1
seaborn==0.12.2
pytest-playwright==0.4.0 